Reorganize Builtins for Options-like behavior
