+++ mathics/builtin/options/__init__.py
@@ -1,0 +1,10 @@
+"""
+Options Management
+
+A number of functions have various options which control the behavior or \
+the default behavior that function. Default options can be queried or set.
+
+<url>
+:WMA link:
+https://reference.wolfram.com/language/guide/OptionsManagement.html</url>
+"""
