+++ mathics/builtin/options/functions.py
@@ -1,133 +1,17 @@
-# -*- coding: utf-8 -*-
-
 """
-Options Management
-
-A number of functions have various options which control the behavior or \
-the default behavior that function. Default options can be queried or set.
-
-<url>
-:WMA link:
-https://reference.wolfram.com/language/guide/OptionsManagement.html</url>
+Setting Up Options for Functions
 """
-from typing import Callable, Optional
 
 from mathics.core.atoms import Integer1, String
-from mathics.core.builtin import Builtin, Predefined, Test, get_option
+from mathics.core.builtin import Builtin, get_option
 from mathics.core.evaluation import Evaluation
 from mathics.core.expression import Expression
 from mathics.core.list import ListExpression
-from mathics.core.parser import parse_builtin_rule
-from mathics.core.symbols import Symbol, SymbolList, ensure_context, strip_context
-from mathics.core.systemsymbols import SymbolDefault, SymbolRule
-from mathics.eval.options import eval_Option_with_names, eval_Options
-from mathics.eval.patterns import Matcher, get_default_value
-
-
-class All(Predefined):
-    """
-    <url>
-    :WMA link:
-    https://reference.wolfram.com/language/ref/All.html</url>
-
-    <dl>
-      <dt>'All'
-      <dd>is an option value for a number of functions indicating to include everything.
-    </dl>
-
-
-    In list functions, it indicates all levels of the list.
-
-    For example, in <url>
-    :Part:
-    /doc/reference-of-built-in-symbols/list-functions/elements-of-lists/part</url>, \
-    'All', extracts into a first column vector the first element of each of the \
-    list elements:
-
-    >> {{1, 3}, {5, 7}}[[All, 1]]
-     = {1, 5}
-
-    While in <url>
-    :Take:
-    /doc/reference-of-built-in-symbols/list-functions/elements-of-lists/part</url>, \
-    'All' extracts as a column matrix the first element as a list for each of the list \
-    elements:
-
-    >> Take[{{1, 3}, {5, 7}}, All, {1}]
-     = {{1}, {5}}
-
-    In <url>
-    :Plot:
-    /doc/reference-of-built-in-symbols/plotting-graphing-and-drawing/general-graphical-plots/plot</url>, \
-    setting the <url>
-    :Mesh:
-/doc/reference-of-built-in-symbols/plotting-graphing-and-drawing/drawing-options-and-option-values/mesh</url> \
-    option to 'All' will show the specific plot points:
-
-    >> Plot[x^2, {x, -1, 1}, MaxRecursion->5, Mesh->All]
-     = -Graphics-
-
-    """
-
-    summary_text = "option value that specify using everything"
-
-
-class Default(Builtin):
-    """
-    <url>
-      :WMA link:
-      https://reference.wolfram.com/language/ref/Default.html</url>
-
-    <dl>
-      <dt>'Default'[$f$]
-      <dd>gives the default value for an omitted parameter of $f$.
-
-      <dt>'Default'[$f$, $k$]
-      <dd>gives the default value for a parameter on the $k$-th position.
-
-      <dt>'Default'[$f$, $k$, $n$]
-      <dd>gives the default value for the $k$-th parameter out of $n$.
-    </dl>
-
-    Assign values to 'Default' to specify default values.
-
-    >> Default[f] = 1
-     = 1
-    >> f[x_.] := x ^ 2
-    >> f[]
-     = 1
-
-    Default values are stored in 'DefaultValues':
-    >> DefaultValues[f]
-     = {HoldPattern[Default[f]] ⧴ 1}
-
-    You can use patterns for $k$ and $n$:
-    >> Default[h, k_, n_] := {k, n}
-    Note that the position of a parameter is relative to the pattern, not the matching expression:
-    >> h[] /. h[___, ___, x_., y_., ___] -> {x, y}
-     = {{3, 5}, {4, 5}}
-    """
-
-    summary_text = "predefined default arguments for a function"
-
-    def eval(self, f, i, evaluation):
-        "Default[f_, i___]"
-
-        i = i.get_sequence()
-        if len(i) > 2:
-            evaluation.message(SymbolDefault, "argb", 1 + len(i), 1, 3)
-            return
-        i = [index.get_int_value() for index in i]
-        for index in i:
-            if index is None or index < 1:
-                evaluation.message(SymbolDefault.name, "intp")
-                return
-        name = f.get_name()
-        if not name:
-            evaluation.message(SymbolDefault.name, "sym", f, 1)
-            return
-        result = get_default_value(name, evaluation, *i)
-        return result
+from mathics.core.symbols import Symbol, ensure_context
+from mathics.core.systemsymbols import SymbolRule
+from mathics.eval.options.function import options_to_rules
+from mathics.eval.options.options import eval_Option_with_names, eval_Options
+from mathics.eval.patterns import Matcher
 
 
 class FilterRules(Builtin):
@@ -169,134 +53,6 @@ class FilterRules(Builtin):
                     yield rule
 
         return ListExpression(*list(matched()))
-
-
-class FormatType(Predefined):
-    """
-    <url>:WMA link:https://reference.wolfram.com/language/ref/FormatType.html</url>
-    <dl>
-      <dt>'FormatType'
-      <dd>is an option for output streams, graphics and functions like 'Text' \
-          that specifies the default format.
-    </dl>
-    """
-
-    messages = {"ftype": "Value of option FormatType -> `` is not valid."}
-    summary_text = "specify the request format"
-
-
-class None_(Predefined):
-    """
-        <url>:WMA link:https://reference.wolfram.com/language/ref/None.html</url>
-
-        <dl>
-          <dt>'None'
-          <dd>is a setting value for many options.
-        </dl>
-
-        Plot3D shows the mesh grid between computed points by default. This the <url>
-        :Mesh:
-/doc/reference-of-built-in-symbols/plotting-graphing-and-drawing/drawing-options-and-option-values/mesh</url> \
-
-        However, you hide the mesh by setting the 'Mesh' option value to 'None':
-
-        >> Plot3D[{x^2 + y^2, -x^2 - y^2}, {x, -2, 2}, {y, -2, 2}, BoxRatios-> Automatic, Mesh->None]
-         = -Graphics3D-
-    """
-
-    name = "None"
-    summary_text = "option value that disables the option"
-
-
-# Has this been removed from WL? I cannot find a WMA link.
-class NotOptionQ(Test):
-    """
-    <url>
-    :WMA link:
-    https://reference.wolfram.com/language/ref/NotOptionQ.html</url>
-
-    <dl>
-      <dt>'NotOptionQ'[$expr$]
-      <dd>returns 'True' if $expr$ does not have the form of a valid \
-          option specification.
-    </dl>
-
-    >> NotOptionQ[x]
-     = True
-    >> NotOptionQ[2]
-     = True
-    >> NotOptionQ["abc"]
-     = True
-
-    >> NotOptionQ[a -> True]
-     = False
-    """
-
-    summary_text = "test whether an expression does not match the form of a valid option specification"
-
-    def test(self, expr) -> bool:
-        if hasattr(expr, "flatten_with_respect_to_head"):
-            expr = expr.flatten_with_respect_to_head(SymbolList)
-        if not expr.has_form("List", None):
-            expr = [expr]
-        else:
-            expr = expr.elements
-        return not all(
-            e.has_form("Rule", None) or e.has_form("RuleDelayed", 2) for e in expr
-        )
-
-
-# Has this been removed from WL? I cannot find a WMA link.
-class OptionQ(Test):
-    """
-    <url>
-    :WMA link:
-    https://reference.wolfram.com/language/ref/OptionQ.html</url>
-
-    <dl>
-      <dt>'OptionQ'[$expr$]
-      <dd>returns 'True' if $expr$ has the form of a valid option \
-         specification.
-    </dl>
-
-    Examples of option specifications:
-    >> OptionQ[a -> True]
-     = True
-    >> OptionQ[a :> True]
-     = True
-    >> OptionQ[{a -> True}]
-     = True
-    >> OptionQ[{a :> True}]
-     = True
-
-    Options lists are flattened when are applied, so
-    >> OptionQ[{a -> True, {b->1, "c"->2}}]
-     = True
-    >> OptionQ[{a -> True, {b->1, c}}]
-     = False
-    >> OptionQ[{a -> True, F[b->1,c->2]}]
-     = False
-
-    'OptionQ' returns 'False' if its argument is not a valid option
-    specification:
-    >> OptionQ[x]
-     = False
-    """
-
-    summary_text = (
-        "test whether an expression matches the form of a valid option specification"
-    )
-
-    def test(self, expr) -> bool:
-        if hasattr(expr, "flatten_with_respect_to_head"):
-            expr = expr.flatten_with_respect_to_head(SymbolList)
-        if not expr.has_form("List", None):
-            expr = [expr]
-        else:
-            expr = expr.elements
-        return all(
-            e.has_form("Rule", None) or e.has_form("RuleDelayed", 2) for e in expr
-        )
 
 
 class Options(Builtin):
@@ -589,44 +345,6 @@ class SetOptions(Builtin):
         return ListExpression(*options_list)
 
 
-def filter_non_default_values(builtin):
-    """
-    Return a filter function that removes those
-    options which have associated their default values.
-    """
-    builtin_options = builtin.options
-    builtin_options = {
-        strip_context(name): parse_builtin_rule(value)
-        for name, value in builtin_options.items()
-    }
-
-    def filter(name, value):
-        name = strip_context(name)
-        if name not in builtin_options:
-            return True
-        if value.sameQ(builtin_options[name]):
-            return False
-        return True
-
-    return filter
-
-
-def filter_from_iterable(elems):
-    """
-    Build a filter function from an iterable.
-    The filter function returns `True` if
-    the name after striping its context is in
-    the interable.
-    """
-
-    def filter(name, value):
-        return strip_context(name) in elems
-
-    return filter
-
-
-def options_to_rules(options, filter: Optional[Callable] = None):
-    items = sorted(options.items())
-    if filter is not None:
-        items = [(name, value) for name, value in items if filter(name, value)]
-    return [Expression(SymbolRule, Symbol(name), value) for name, value in items]
+## FIXME add:
+## AbsoluteOptions
+## ArgumentOptions
